# docs/api/swagger-na0wdotzh8.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Sample App API Gateway",
    "version" : "2024-09-11T17:18:38Z",
    "title" : "imdbapp-backend-api"
  },
  "host" : "na0wdotzh8.execute-api.us-east-1.amazonaws.com",
  "basePath" : "/v1",
  "schemes" : [ "https" ],
  "paths" : {
...
    "/test/secureIAM" : {
      "get" : {
        "responses" : { },
        "security" : [ {
          "sigv4" : [ ]
        } ],
        "x-postman-meta" : {
          "helperAttributes" : {
            "saveHelper" : true,
            "secretKey" : "{{aws_secret_access_key}}",
            "accessKey" : "{{aws_access_key_id}}"
          },
          "headers" : "Content-Type: application/json\n",
          "currentHelper" : "awsSigV4"
        }
      },
      "options" : {
        "consumes" : [ "application/json" ],
        "responses" : {
          "204" : {
            "description" : "204 response",
            "headers" : {
              "Access-Control-Allow-Origin" : {
                "type" : "string"
              },
              "Access-Control-Allow-Methods" : {
                "type" : "string"
              },
              "Access-Control-Allow-Credentials" : {
                "type" : "string"
              },
              "Access-Control-Allow-Headers" : {
                "type" : "string"
              }
            }
          }
        }
      }
    },
    "/test/secureIAMAdmin" : {
      "get" : {
        "responses" : { },
        "security" : [ {
          "sigv4" : [ ]
        } ],
        "x-postman-meta" : {
          "helperAttributes" : {
            "saveHelper" : true,
            "secretKey" : "{{aws_secret_access_key}}",
            "accessKey" : "{{aws_access_key_id}}"
          },
          "headers" : "Content-Type: application/json\n",
          "currentHelper" : "awsSigV4"
        }
      },
      "options" : {
        "consumes" : [ "application/json" ],
        "responses" : {
          "204" : {
            "description" : "204 response",
            "headers" : {
              "Access-Control-Allow-Origin" : {
                "type" : "string"
              },
              "Access-Control-Allow-Methods" : {
                "type" : "string"
              },
              "Access-Control-Allow-Credentials" : {
                "type" : "string"
              },
              "Access-Control-Allow-Headers" : {
                "type" : "string"
              }
            }
          }
        }
      }
    },
    "/test/secureIAMUser" : {
      "get" : {
        "responses" : { },
        "security" : [ {
          "sigv4" : [ ]
        } ],
        "x-postman-meta" : {
          "helperAttributes" : {
            "saveHelper" : true,
            "secretKey" : "{{aws_secret_access_key}}",
            "accessKey" : "{{aws_access_key_id}}"
          },
          "headers" : "Content-Type: application/json\n",
          "currentHelper" : "awsSigV4"
        }
      },
      "options" : {
        "consumes" : [ "application/json" ],
        "responses" : {
          "204" : {
            "description" : "204 response",
            "headers" : {
              "Access-Control-Allow-Origin" : {
                "type" : "string"
              },
              "Access-Control-Allow-Methods" : {
                "type" : "string"
              },
              "Access-Control-Allow-Credentials" : {
                "type" : "string"
              },
              "Access-Control-Allow-Headers" : {
                "type" : "string"
              }
            }
          }
        }
      }
    },
...
  },
  "securityDefinitions" : {
    "sigv4" : {
      "type" : "apiKey",
      "name" : "Authorization",
      "in" : "header",
      "x-amazon-apigateway-authtype" : "awsSigv4"
    }
  }
}